code sync

+++ controllers/hotel.js
@@ -3,20 +3,20 @@ import Order from "../models/order";
 import fs from "fs";
 
 export const bookHotel = async (req, res) => {
-  //   console.log("req.fields", req.fields);
-  //   console.log("req.files", req.files);
   try {
-    // let fields = req.fields;
-    const { hotelId, amount, bookingDetails}  = req.body;
+    const { hotelId, amount, bookingDetails } = req.body;
     let order = new Order();
     order.orderedBy = req.user._id;
     order.hotel = hotelId;
     order.bookingDetails = JSON.parse(bookingDetails);
+    order.bed = order.bookingDetails.bed
+    order.from = order.bookingDetails.from
+    order.to = order.bookingDetails.to
     order.session = {
-      "payment_intent":"PID",
-      "payment_status":"PAID",
-      "currency":"INR",
-      "amount_total":amount
+      "payment_intent": "PID",
+      "payment_status": "PAID",
+      "currency": "INR",
+      "amount_total": amount
     };
     order.save((err, result) => {
       if (err) {
@@ -34,12 +34,9 @@ export const bookHotel = async (req, res) => {
 };
 
 export const create = async (req, res) => {
-  //   console.log("req.fields", req.fields);
-  //   console.log("req.files", req.files);
   try {
     let fields = req.fields;
     let files = req.files;
-
     let hotel = new Hotel(fields);
     hotel.postedBy = req.user._id;
     // handle image
@@ -64,13 +61,11 @@ export const create = async (req, res) => {
 };
 
 export const hotels = async (req, res) => {
-  // let all = await Hotel.find({ from: { $gte: new Date() } })
   let all = await Hotel.find({})
     .limit(24)
     .select("-image.data")
     .populate("postedBy", "_id name")
     .exec();
-  // console.log(all);
   res.json(all);
 };
 
@@ -92,6 +87,11 @@ export const sellerHotels = async (req, res) => {
 };
 
 export const remove = async (req, res) => {
+  // First find if hotel has orders history : can't delete
+  const ordersCount = await Order.countDocuments({ hotel: req.params.hotelId }).exec();
+  if (ordersCount) {
+    return res.status(400).send("Can't delete, this has orders history");
+  }
   let removed = await Hotel.findByIdAndDelete(req.params.hotelId)
     .select("-image.data")
     .exec();
@@ -103,7 +103,6 @@ export const read = async (req, res) => {
     .populate("postedBy", "_id name")
     .select("-image.data")
     .exec();
-  // console.log("SINGLE HOTEL", hotel);
   res.json(hotel);
 };
 
@@ -135,7 +134,7 @@ export const update = async (req, res) => {
 
 export const userHotelBookings = async (req, res) => {
   const all = await Order.find({ orderedBy: req.user._id })
-    .select("session bookingDetails")
+    .select("session bed from to bookingDetails")
     .populate("hotel", "-image.data")
     .populate("orderedBy", "_id name")
     .exec();
@@ -160,47 +159,131 @@ export const isAlreadyBooked = async (req, res) => {
 
 export const searchListings = async (req, res) => {
   const { location, date, bed } = req.body;
-  console.log(location, date, bed);
-  // console.log(date);
+  let where = { location };
 
-  let where = {location};
-  // {
-  //   from: { $gte: new Date(fromDate[0]) },
-  //   location,
-  // }
-
-  if(date) {
+  if (date) {
     const fromDate = date.split(",");
-    if(fromDate && fromDate[0]) {
-      where = {...where, from: { $lte: new Date(fromDate[0]) }}
+    if (fromDate && fromDate[0]) {
+      where = { ...where, from: { $lte: new Date(fromDate[0]) } }
     }
 
-    if(fromDate && fromDate[1]) {
-      where = {...where, to: { $gte: new Date(fromDate[1]) }}
+    if (fromDate && fromDate[1]) {
+      where = { ...where, to: { $gte: new Date(fromDate[1]) } }
     }
+  }
+
+  if (bed) {
+    where = { ...where, bed: { $gte: bed } }
+  }
+
+  const fromDate = date.split(",");
+  let result = await Hotel.find(where)
+    .select("-image.data")
+    .exec();
+  res.json(result);
+};
+
+export const cancelBooking = async (req, res) => {
+  let removed = await Order.findByIdAndDelete(req.params.orderId)
+    .exec();
+  res.json(removed);
+};
+
+export const checkHotelAvailability = async (req, res) => {
+  const { from, to, bed, hotelId } = req.body;
+  let hotelObj = await Hotel.findById(hotelId).exec();
+  let bookedBeds = 0;
+  let where = { hotel: hotelId };
+
+  if (from && to) { // CASE 1
+    where = { hotel: hotelId };
+    where = { ...where, from: { $lte: new Date(from) } }
+    where = { ...where, to: { $gte: new Date(from) } }
+  }
+
+  let result1 = await Order.find(where).select("from to bed").exec();
+  where = {};
+
+  if (from && to) { // CASE 2
+    where = { hotel: hotelId };
+    where = { ...where, from: { $gte: new Date(from) } }
+    where = { ...where, to: { $lte: new Date(to) } }
+  }
+
+  let result2 = await Order.find(where).select("from to bed").exec();
+  where = {};
+ 
+
+  if (from && to) { // CASE 3
+    where = { hotel: hotelId };
+    where = { ...where, from: { $lte: new Date(to) } }
+    where = { ...where, to: { $gte: new Date(to) } }
+  }
+
+  let result3 = await Order.find(where).select("from to bed").exec();
+  where = {};
+
+  if (from && to) { // CASE 4
+    where = { hotel: hotelId };
+    where = { ...where, from: { $lte: new Date(from) } }
+    where = { ...where, to: { $gte: new Date(to) } }
+  }
+
+  let result4 = await Order.find(where).select("from to bed").exec();
+  where = {};
+ 
+
+  let result = await Order.find(where).select("from to bed").exec();
+  // console.log("result1",result1)
+  // console.log("result2",result2)
+  // console.log("result3",result3)
+  // console.log("result4",result4)
+
+  const finalResult = [
+    ...result1,...result2,...result3,...result4
+  ]
+
+  if(finalResult.length) {
+    for (let i = 0; i < finalResult.length; i++) {
+      bookedBeds += finalResult[i].bed;
+    }
+  } else {
+    console.log("yes i can book")
+  }
+
+  if(!bookedBeds) {
+    return res.json({
+      canBook: true
+    });
   }
 
   
 
-  if(bed){
-    where = {...where, bed: { $gte: bed }}
+  const remaingBeds = hotelObj.bed - bookedBeds;
+  console.log(remaingBeds, hotelObj.bed , bookedBeds,parseInt(bed))
+  if(parseInt(bed) <= remaingBeds) {
+    return res.json({
+      canBook: true
+    });
   }
 
-  const fromDate = date.split(",");
-  // console.log(fromDate[0]);
-  let result = await Hotel.find(where)
-    .select("-image.data")
-    .exec();
-  // console.log("SEARCH LISTINGS", result);
-  res.json(result);
+  return res.json({
+    canBook: false
+  });
 };
 
-/**
- * if you want to be more specific
- let result = await Listing.find({
-  from: { $gte: new Date() },
-  to: { $lte: to },
-  location,
-  bed,
-})
- */
+export const getHotelBookings = async (req, res) => {
+  let all = await Order.find({ hotel: req.params.hotelId })
+    .limit(24)
+    .select("bed from to bookingDetails session")
+    .populate("orderedBy", "name")
+    .exec();
+  res.json(all);
+};
+
+
+export const getCompletedOrdersCount = async (req, res) => {
+  let ordersCount = await Order.countDocuments({ to: { $lt: new Date() }, orderedBy: req.params.userlId }).exec();
+  res.json(ordersCount)
+};
+
